# apps/lumistripe-web/frontend/openapi.json
{
  "components": {
    "schemas": {
      "AccessStatus": {
        "properties": {
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          }
        },
        "required": [
          "required",
          "authenticated"
        ],
        "title": "AccessStatus",
        "type": "object"
      },
...
      "AnimationRequest": {
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "stripe_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "AnimationRequest",
        "type": "object"
      },
...
      "BlackoutRequest": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "stripe_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Id"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "BlackoutRequest",
        "type": "object"
      },
...
      "BrightnessRequest": {
        "properties": {
          "brightness": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Brightness",
            "type": "number"
          },
          "stripe_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Id"
          }
        },
        "required": [
          "brightness"
        ],
        "title": "BrightnessRequest",
        "type": "object"
      },
      "CalibrationFinishRequest": {
        "properties": {
          "save": {
            "title": "Save",
            "type": "boolean"
          }
        },
        "required": [
          "save"
        ],
        "title": "CalibrationFinishRequest",
        "type": "object"
      },
      "CalibrationSessionResponse": {
        "properties": {
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/DashboardState"
          }
        },
        "required": [
          "session_id",
          "state"
        ],
        "title": "CalibrationSessionResponse",
        "type": "object"
      },
      "CalibrationStartRequest": {
        "properties": {
          "output_index": {
            "minimum": 0.0,
            "title": "Output Index",
            "type": "integer"
          }
        },
        "required": [
          "output_index"
        ],
        "title": "CalibrationStartRequest",
        "type": "object"
      },
      "CalibrationStatus": {
        "properties": {
          "active": {
            "default": false,
            "title": "Active",
            "type": "boolean"
          },
          "expires_in_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Seconds"
          },
          "output_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Index"
          },
          "pattern": {
            "anyOf": [
              {
                "enum": [
                  "white",
                  "red",
                  "green",
                  "blue"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pattern"
          }
        },
        "title": "CalibrationStatus",
        "type": "object"
      },
      "CalibrationUpdateRequest": {
        "properties": {
          "blue": {
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "Blue",
            "type": "integer"
          },
          "green": {
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "Green",
            "type": "integer"
          },
          "pattern": {
            "enum": [
              "white",
              "red",
              "green",
              "blue"
            ],
            "title": "Pattern",
            "type": "string"
          },
          "red": {
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "Red",
            "type": "integer"
          }
        },
        "required": [
          "red",
          "green",
          "blue",
          "pattern"
        ],
        "title": "CalibrationUpdateRequest",
        "type": "object"
      },
...
      "DashboardState": {
        "properties": {
          "active_effects": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Active Effects",
            "type": "array"
          },
          "animation": {
            "default": "",
            "title": "Animation",
            "type": "string"
          },
          "application_version": {
            "default": "development",
            "title": "Application Version",
            "type": "string"
          },
          "audio_callback_age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Callback Age Seconds"
          },
          "audio_frame_age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Frame Age Seconds"
          },
          "audio_health": {
            "default": "inactive",
            "title": "Audio Health",
            "type": "string"
          },
          "audio_status": {
            "default": "No audio source active.",
            "title": "Audio Status",
            "type": "string"
          },
          "blackout": {
            "default": false,
            "title": "Blackout",
            "type": "boolean"
          },
          "bpm": {
            "default": 0.0,
            "title": "Bpm",
            "type": "number"
          },
          "brightness": {
            "default": 1.0,
            "title": "Brightness",
            "type": "number"
          },
          "calibration": {
            "$ref": "#/components/schemas/CalibrationStatus",
            "default": {
              "active": false
            }
          },
          "color_corrections": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ColorCorrectionProfile"
            },
            "title": "Color Corrections",
            "type": "array"
          },
          "command_queue_depth": {
            "default": 0,
            "title": "Command Queue Depth",
            "type": "integer"
          },
          "diagnostic_issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DiagnosticIssue"
            },
            "title": "Diagnostic Issues",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "frame_rate": {
            "default": 0.0,
            "title": "Frame Rate",
            "type": "number"
          },
          "last_output_age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Output Age Seconds"
          },
          "last_output_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Output At"
          },
          "missed_frame_count": {
            "default": 0,
            "title": "Missed Frame Count",
            "type": "integer"
          },
          "mode": {
            "$ref": "#/components/schemas/PlaybackMode",
            "default": "static"
          },
          "music_active": {
            "default": false,
            "title": "Music Active",
            "type": "boolean"
          },
          "music_gate": {
            "default": "calm",
            "title": "Music Gate",
            "type": "string"
          },
          "music_recognition_enabled": {
            "default": true,
            "title": "Music Recognition Enabled",
            "type": "boolean"
          },
          "output_backend": {
            "default": "simulation",
            "title": "Output Backend",
            "type": "string"
          },
          "output_devices": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Output Devices",
            "type": "array"
          },
          "power_budget": {
            "$ref": "#/components/schemas/PowerBudgetState",
            "default": {
              "applied_scale": 1.0,
              "enabled": false,
              "estimated_watts": 0.0,
              "outputs": []
            }
          },
          "revision": {
            "default": 0,
            "title": "Revision",
            "type": "integer"
          },
          "running": {
            "default": false,
            "title": "Running",
            "type": "boolean"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeKind"
          },
          "solid_color": {
            "default": "#7C3AED",
            "title": "Solid Color",
            "type": "string"
          },
          "spi_speed_hz": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spi Speed Hz"
          },
          "stripe_playback": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/StripePlaybackState"
            },
            "title": "Stripe Playback",
            "type": "array"
          },
          "stripe_topology": {
            "$ref": "#/components/schemas/StripeTopology",
            "default": {
              "layout": "mirrored",
              "outputs": [],
              "power_budget_enabled": false
            }
          },
          "uptime_seconds": {
            "default": 0.0,
            "title": "Uptime Seconds",
            "type": "number"
          },
          "worker_heartbeat_age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Heartbeat Age Seconds"
          }
        },
        "required": [
          "runtime"
        ],
        "title": "DashboardState",
        "type": "object"
      },
      "DiagnosticIssue": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "severity",
          "title",
          "message",
          "action"
        ],
        "title": "DiagnosticIssue",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ModeRequest": {
        "properties": {
          "color": {
            "anyOf": [
              {
                "pattern": "^#[0-9A-Fa-f]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "mode": {
            "$ref": "#/components/schemas/PlaybackMode"
          },
          "music_recognition_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Music Recognition Enabled"
          },
          "stripe_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Id"
          }
        },
        "required": [
          "mode"
        ],
        "title": "ModeRequest",
        "type": "object"
      },
      "PairingRequest": {
        "properties": {
          "code": {
            "pattern": "^[0-9]{4}$",
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "PairingRequest",
        "type": "object"
      },
      "PlaybackMode": {
        "enum": [
          "solid",
          "static",
          "cycling",
          "dynamic"
        ],
        "title": "PlaybackMode",
        "type": "string"
      },
...
      "RuntimeKind": {
        "enum": [
          "simulation",
          "hardware"
        ],
        "title": "RuntimeKind",
        "type": "string"
      },
...
      "StripeTestRequest": {
        "properties": {
          "pattern": {
            "default": "identify",
            "enum": [
              "identify",
              "red",
              "green",
              "blue",
              "white"
            ],
            "title": "Pattern",
            "type": "string"
          },
          "stripe_id": {
            "title": "Stripe Id",
            "type": "string"
          },
          "topology": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StripeTopologyRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "stripe_id"
        ],
        "title": "StripeTestRequest",
        "type": "object"
      },
      "StripeTopology": {
        "properties": {
          "layout": {
            "default": "mirrored",
            "enum": [
              "mirrored",
              "continuous",
              "independent"
            ],
            "title": "Layout",
            "type": "string"
          },
          "outputs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/StripeOutputConfig"
            },
            "maxItems": 2,
            "title": "Outputs",
            "type": "array"
          },
          "power_budget_enabled": {
            "default": false,
            "title": "Power Budget Enabled",
            "type": "boolean"
          },
          "power_budget_watts": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Budget Watts"
          }
        },
        "title": "StripeTopology",
        "type": "object"
      },
      "StripeTopologyRequest": {
        "properties": {
          "layout": {
            "enum": [
              "mirrored",
              "continuous",
              "independent"
            ],
            "title": "Layout",
            "type": "string"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/StripeOutputConfig"
            },
            "maxItems": 2,
            "title": "Outputs",
            "type": "array"
          },
          "power_budget_enabled": {
            "default": false,
            "title": "Power Budget Enabled",
            "type": "boolean"
          },
          "power_budget_watts": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Budget Watts"
          }
        },
        "required": [
          "layout",
          "outputs"
        ],
        "title": "StripeTopologyRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LumiStripe",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/animation": {
      "put": {
        "operationId": "select_animation_api_animation_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Animation"
      }
    },
...
    "/api/auth/pair": {
      "post": {
        "operationId": "pair_api_auth_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pair"
      }
    },
...
    "/api/blackout": {
      "post": {
        "operationId": "set_blackout_api_blackout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlackoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Blackout"
      }
    },
    "/api/brightness": {
      "put": {
        "operationId": "set_brightness_api_brightness_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrightnessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Brightness"
      }
    },
    "/api/calibration/session": {
      "post": {
        "operationId": "start_calibration_api_calibration_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Calibration"
      }
    },
    "/api/calibration/session/{session_id}": {
      "put": {
        "operationId": "update_calibration_api_calibration_session__session_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Calibration"
      }
    },
    "/api/calibration/session/{session_id}/finish": {
      "post": {
        "operationId": "finish_calibration_api_calibration_session__session_id__finish_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finish Calibration"
      }
    },
...
    "/api/mode": {
      "put": {
        "operationId": "set_mode_api_mode_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Mode"
      }
    },
...
    "/api/stripes": {
      "get": {
        "operationId": "stripe_topology_api_stripes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StripeTopology"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Stripe Topology"
      },
      "put": {
        "operationId": "update_stripe_topology_api_stripes_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StripeTopologyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Stripe Topology"
      }
    },
    "/api/stripes/test": {
      "post": {
        "operationId": "test_stripe_api_stripes_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StripeTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Stripe"
      }
    }
  }
}